fix: `PostServiceValidator` moved into `validator` package
