+++ src/main/java/faang/school/postservice/service/post/PostService.java
@@ -10,15 +10,12 @@ import faang.school.postservice.dto.resource.UpdatableResourceDto;
 import faang.school.postservice.exception.messages.ValidationExceptionMessage;
 import faang.school.postservice.exception.post.UnexistentPostException;
 import faang.school.postservice.exception.validation.DataValidationException;
-import faang.school.postservice.mapper.post.MediaMapper;
 import faang.school.postservice.mapper.post.PostMapper;
 import faang.school.postservice.mapper.post.ResourceMapper;
 import faang.school.postservice.model.Post;
-import faang.school.postservice.model.Resource;
 import faang.school.postservice.repository.PostRepository;
-import faang.school.postservice.repository.ResourceRepository;
 import faang.school.postservice.service.post.command.UpdatePostResourceCommand;
-import faang.school.postservice.service.post.validator.PostServiceValidator;
+import faang.school.postservice.validator.post.PostServiceValidator;
 import faang.school.postservice.service.resource.ResourceService;
 import jakarta.validation.constraints.NotNull;
 import lombok.RequiredArgsConstructor;
